Updates to dunder_main
